requirements.txt
torch~=2.6.0
numpy~=2.2.4
gymnasium~=1.1.1
pandas~=2.2.3
matplotlib~=3.10.1
pillow~=11.2.1